- Fixes a bug for updating a contact's email address in HubSpot. -- In order to effectively change an email address, we must account for the possibility that a user may change their email address more than once inside a 6-hour period (how often the sync process runs); perhaps they change their mind about which email address to use OR they're correcting a typo after changing, etc. This means that strictly for the EMAIL contact attribute we must explicitly find the "dbo.vw_crds_audit_log.PreviousValue" of the initial email change and the "dbo.vw_crds_audit_log.NewValue" of the most recent email change. For instance...
PreviousValue     |     NewValue     |     Updated
0@0.com                 1@1.com            2018-06-01 08:00:00 AM
1@1.com                 2@2.com            2018-06-01 09:00:00 AM
2@2.com                 3@3.com            2018-06-01 10:00:00 AM
3@3.com                 4@4.com            2018-06-01 11:00:00 AM

In this scenario, we run the process at 8:05AM, meaning HubSpot now knows about 1@1.com as the current email address. since then, the email address was changed 3 times inside of a window of time before the sync process runs again. So we'll need to query the audit log to get the email address before it changed and query it again for the last time it changed, meaning 1@1.com in row 2 will be found for the current email address and 4@4.com will be found for the most recent change.